Refactor move client files
To reorganize the project structure
