+++ client/src/common/utils.ts
